More shop features and fixes

+++ src/game_state.hpp
@@ -5,5 +5,5 @@
 
 struct GameState {
 	double money;
-	std::map<std::wstring, double> upgrades;
+	std::map<std::wstring, float> upgrades;
 };
